Creación de función de verificación
