+++ src/routes/teachers.routes.js
@@ -9,5 +9,6 @@ router.get("/:id", teachersController.getTeacherByIdHandler);
 router.delete("/:id", teachersController.deleteTeacherByIdHandler);
 router.put("/:id", teachersController.updateTeacherHandler);
 
+router.get("/:email",teachersController.checkTeacherEmailExists);
 
 export default router;
